requirements.txt
asyncio
pyrofork
tgcrypto
pymongo
python-dotenv
requests